Bug 2022925: Don't cast denominator to int32 when lowering LDivConstantI64 for x64. r=spidermonkey-reviewers,jandem
Also added more test coverage for signed and unsigned division and modulus operations.

Differential Revision: https://phabricator.services.mozilla.com/D287568

UltraBlame original commit: 80e42d0ac16dc0a8c6c8e1bd763f8a2bf04c53fe

+++ js/src/jit/x64/CodeGenerator-x64.cpp
@@ -486,7 +486,7 @@ static void Divide64WithConstant(MacroAssembler& masm, LDivOrMod* ins) {
 }
 
 void CodeGenerator::visitDivConstantI64(LDivConstantI64* ins) {
-  int32_t d = ins->denominator();
+  int64_t d = ins->denominator();
 
   
   MOZ_ASSERT(ToRegister(ins->output()) == rdx);
